+++ .gitignore
@@ -8,6 +8,7 @@ __pycache__/
 /log.jsonl
 /packages/vscode/bin/
 /run/
+/stats/
 /tapenade_3.16.tar
 /tapenade_3.16/
 /target/
