+++ crates/gradbench/Cargo.toml
@@ -14,6 +14,7 @@ clap = { version = "4", features = ["derive"] }
 colored = "2"
 ctrlc = "3"
 nix = { version = "0.29", features = ["process", "signal"] }
+plotters = "0.3"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 strum = { version = "0.27", features = ["derive"] }
