+++ crates/gradbench/src/main.rs
@@ -1,10 +1,10 @@
 mod intermediary;
 mod protocol;
+mod stats;
 
 use std::{
     collections::HashSet,
-    env, fs,
-    io::{self, BufRead},
+    env, fs, io,
     path::{Path, PathBuf},
     process::{Command, ExitCode, ExitStatus, Output, Stdio},
     str::FromStr,
@@ -12,8 +12,8 @@ use std::{
 };
 
 use clap::{Parser, Subcommand};
-use protocol::Message;
 use serde::Serialize;
+use stats::StatsMetadata;
 use strum::{EnumIter, EnumString, IntoStaticStr};
 
 /// CLI utilities for GradBench, a benchmark suite for differentiable programming across languages
@@ -171,24 +171,25 @@ enum RepoCommands {
     /// sign. No extra whitespace is printed, because GitHub Actions seems to be sensitive to that.
     Matrix,
 
-    /// Generate summary data files in a directory containing log files.
+    /// Generate summary data files and plots from a directory containing log files.
     ///
-    /// The directory should first contain a `run-<EVAL>-<TOOL>/log.jsonl` file for each `<EVAL>`
-    /// under `evals` and each `<TOOL>` under `tools`.
-    ///
-    /// Used in the nightly builds, producing files that can be easily downloaded by JavaScript on
-    /// the GradBench website to generate tables and charts.
-    Summarize {
-        /// The directory to work in
-        dir: PathBuf,
+    /// The directory should contain a `run-<EVAL>-<TOOL>/log.jsonl` file for each `<EVAL>` under
+    /// `evals` and each `<TOOL>` under `tools`.
+    Stats {
+        /// The directory containing log files
+        input: PathBuf,
+
+        /// The directory to create
+        #[clap(short, long)]
+        output: PathBuf,
 
         /// The current date
         #[clap(long)]
-        date: String,
+        date: Option<String>,
 
-        /// The Git commit SHA from the `main` branch
+        /// The source Git commit SHA
         #[clap(long)]
-        commit: String,
+        commit: Option<String>,
     },
 }
 
@@ -442,77 +443,6 @@ struct RunEntry<'a> {
     outcome: &'static str,
 }
 
-/// The status from running a tool on an eval.
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "snake_case")]
-enum Status {
-    /// The tool is not implemented for the eval.
-    Unimplemented,
-
-    /// The tool returned invalid results for the eval.
-    Incorrect,
-
-    /// The tool returned correct results for the eval.
-    Correct,
-}
-
-/// Read a log file and summarize the results.
-fn summarize(path: &Path) -> Option<Status> {
-    let mut message: Option<Message> = None;
-    for res in io::BufReader::new(fs::File::open(path).ok()?).lines() {
-        let mut line: serde_json::Value = serde_json::from_str(&res.ok()?).ok()?;
-        if let Some(response) = line.get("response") {
-            if let Message::Define { .. } = message.take()? {
-                if !response.get("success")?.as_bool()? {
-                    return Some(Status::Unimplemented);
-                }
-            }
-        }
-        if let Some(msg) = line.get_mut("message") {
-            message = Some(serde_json::from_value(msg.take()).ok()?);
-            if let Some(Message::Analysis { valid, .. }) = &message {
-                if !valid {
-                    return Some(Status::Incorrect);
-                }
-            }
-        }
-    }
-    Some(Status::Correct)
-}
-
-/// A cell in a table of summary data.
-#[derive(Debug, Serialize)]
-struct Col<'a> {
-    /// The name of the tool for this column.
-    tool: &'a str,
-
-    /// The status of the tool for this eval.
-    status: Status,
-}
-
-/// A row in a table of summary data.
-#[derive(Debug, Serialize)]
-struct Row<'a> {
-    /// The name of the eval for this row.
-    eval: &'a str,
-
-    /// The status of each tool for this eval.
-    tools: Vec<Col<'a>>,
-}
-
-/// Summary data to be written to a `summary.json` file by the `repo summarize` subcommand.
-#[derive(Debug, Serialize)]
-struct Summary<'a> {
-    /// The current date.
-    date: String,
-
-    /// The Git commit SHA from the `main` branch.
-    commit: String,
-
-    /// The table of summary data.
-    table: Vec<Row<'a>>,
-}
-
 /// Run the GradBench CLI, returning a `Result`.
 fn cli() -> Result<(), ExitCode> {
     match Cli::parse().command {
@@ -644,36 +574,16 @@ fn cli() -> Result<(), ExitCode> {
                     github_output("run", run)?;
                     Ok(())
                 }
-                RepoCommands::Summarize { dir, date, commit } => {
-                    let mut table = Vec::new();
-                    let mut evals = ls("evals")?;
-                    evals.sort();
-                    let mut tools = ls("tools")?;
-                    tools.sort();
-                    for eval in &evals {
-                        let mut row = Vec::new();
-                        for tool in &tools {
-                            let path = dir.join(format!("run-{eval}-{tool}/log.jsonl"));
-                            let status = summarize(&path).unwrap_or(Status::Unimplemented);
-                            row.push(Col { tool, status });
-                        }
-                        table.push(Row { eval, tools: row });
-                    }
-                    let summary = Summary {
-                        date,
-                        commit,
-                        table,
-                    };
-                    let output = dir.join("summary.json");
-                    let file = fs::File::create(&output).map_err(|err| {
-                        eprintln!("error creating summary file {output:?}: {err}");
+                RepoCommands::Stats {
+                    input,
+                    output,
+                    date,
+                    commit,
+                } => {
+                    stats::generate(input, output, StatsMetadata { date, commit }).map_err(|err| {
+                        eprintln!("{err:#}");
                         ExitCode::FAILURE
-                    })?;
-                    serde_json::to_writer(file, &summary).map_err(|err| {
-                        eprintln!("error serializing summary table: {err}");
-                        ExitCode::FAILURE
-                    })?;
-                    Ok(())
+                    })
                 }
             }
         }
